# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for go-api.",
        "title": "go-api API",
        "contact": {
            "name": "Lewis",
            "email": "lxx0103@yahoo.com"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
...
        "/historys": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "历史记录管理"
                ],
                "summary": "历史记录列表",
                "operationId": "701",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页行数（5/10/15/20）",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "相关ID",
                        "name": "reference_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "历史类型",
                        "name": "history_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/common.HistoryResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
...
        "/nextnumber": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通用"
                ],
                "summary": "获取下一个编码",
                "operationId": "702",
                "parameters": [
                    {
                        "type": "string",
                        "description": "编码类型",
                        "name": "number_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.HistoryResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "history_by": {
                    "type": "string"
                },
                "history_id": {
                    "type": "string"
                },
                "history_time": {
                    "type": "string"
                },
                "history_type": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ListRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {},
                "page_id": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "response.SuccessRes": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
    }
}